Fixed the Camera, its ready to go.
For the most part the camera is ready to go. The only thing that's not
in yet that I can think of is that when the player enters the
TriggerNodes it could calculate their direction and modify the variables
accordingly. Maybe I'll do that after we get everything else in.